+++ packages/sdk/src/types.ts
@@ -1,7 +1,7 @@
 /**
  * @module Common GRPC-Client
  */
-import type { Upward } from './grpc/index.js';
+import type { Known, Upward } from './grpc/index.js';
 import type { Cbor, TokenId } from './plt/index.js';
 import type { TokenAccountInfo } from './plt/types.js';
 import type * as AccountAddress from './types/AccountAddress.js';
@@ -515,7 +515,16 @@ interface AuthorizationsCommon {
     electionDifficulty: Authorization;
     addAnonymityRevoker: Authorization;
     addIdentityProvider: Authorization;
-    keys: VerifyKey[];
+    /**
+     * The authorization keys.
+     *
+     * **Please note**, these can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     *
+     * In case this is used as part of a transaction sent to the node, none of the values contained can be `null`,
+     * as this will cause the transation to fail.
+     */
+    keys: Upward<VerifyKey>[];
 }
 
 /**
@@ -539,7 +548,13 @@ export interface AuthorizationsV1 extends AuthorizationsCommon {
 export type Authorizations = AuthorizationsV0 | AuthorizationsV1;
 
 export interface KeysWithThreshold {
-    keys: VerifyKey[];
+    /**
+     * The authorization keys.
+     *
+     * **Please note**, these can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     */
+    keys: Upward<VerifyKey>[];
     threshold: number;
 }
 
@@ -798,7 +813,16 @@ export interface VerifyKey {
 }
 
 export interface CredentialPublicKeys {
-    keys: Record<number, VerifyKey>;
+    /**
+     * keys for the credential
+     *
+     * **Please note**, these can possibly be unknown if the SDK is not fully compatible with the Concordium
+     * node queried, in which case `null` is returned.
+     *
+     * In case this is used as part of a transaction sent to the node, none of the values contained can be `null`,
+     * as this will cause the transation to fail.
+     */
+    keys: Record<number, Upward<VerifyKey>>;
     threshold: number;
 }
 
@@ -1661,11 +1685,10 @@ interface CdiRandomness {
     randomness: CommitmentsRandomness;
 }
 
-// TODO Should we rename this, As it is not actually the transaction that is sent to the node. (Note that this would be a breaking change)
-export type CredentialDeploymentTransaction = CredentialDeploymentDetails & CdiRandomness;
+export type CredentialDeploymentPayload = CredentialDeploymentDetails & CdiRandomness;
 /** Internal type used when building credentials */
 export type UnsignedCdiWithRandomness = {
-    unsignedCdi: UnsignedCredentialDeploymentInformation;
+    unsignedCdi: Known<UnsignedCredentialDeploymentInformation>;
 } & CdiRandomness;
 
 export interface CredentialDeploymentInfo extends CredentialDeploymentValues {